Testing for existence before deleting file and adding a message about deleting output files.

+++ generate-numbers.py
@@ -2,7 +2,8 @@
 import random
 import os
 fn = 'rns'
-os.remove(fn)
+if os.path.exists(fn):
+  os.remove(fn)
 a = {}
 f = open('rns', 'w')
 for i in range(1,1000000):
